Factor out abstract geo helper methods to utility Module
