+++ lib/stanford-mods.rb
@@ -2,6 +2,7 @@ require 'mods'
 require 'stanford-mods/date_parsing'
 require 'stanford-mods/coordinate'
 require 'stanford-mods/geo_spatial'
+require 'stanford-mods/geo_utils'
 require 'stanford-mods/imprint'
 require 'stanford-mods/name'
 require 'stanford-mods/origin_info'
