+++ lib/stanford-mods/coordinate.rb
@@ -4,26 +4,28 @@ module Stanford
     ##
     # Geospatial coordinate parsing
     class Coordinate
+      require 'stanford-mods/geo_utils'
+      include ::Stanford::Mods::GeoUtils
+
       attr_reader :value
 
       def initialize(value)
         @value = value
       end
 
-      # Convert the coordinate to a WKT/CQL ENVELOPE representation
+      # @return [String] the coordinate in WKT/CQL ENVELOPE representation
       def as_envelope
         return unless valid?
-
         "ENVELOPE(#{bounds[:min_x]}, #{bounds[:max_x]}, #{bounds[:max_y]}, #{bounds[:min_y]})"
       end
 
-      # Convert the coordinate to a Solr 4.x bbox-format representation
+      # @return [String] the coordinate in Solr 4.x+ bbox-format representation
       def as_bbox
         return unless valid?
-
         "#{bounds[:min_x]} #{bounds[:min_y]} #{bounds[:max_x]} #{bounds[:max_y]}"
       end
 
+      # @return [Boolean] true iff the coordinates are geographically valid
       def valid?
         return false if bounds.empty?
 
@@ -40,7 +42,7 @@ module Stanford
 
       def bounds
         @bounds ||= begin
-          matches = coord.match %r{\A(?<lat>[EW].+-+.+)\s*/\s*(?<lng>[NS].+-+.+)\Z}
+          matches = cleaner_coordinate(value).match %r{\A(?<lat>[EW].+-+.+)\s*/\s*(?<lng>[NS].+-+.+)\Z}
 
           if matches
             min_x, max_x = matches['lat'].split(/-+/).map { |x| coord_to_decimal(x) }.minmax
@@ -54,27 +56,7 @@ module Stanford
       end
 
       def coord
-        matches = value.match(/^\(?([^)]+)\)?\.?$/)
-
-        if matches
-          matches[1]
-        else
-          value
-        end
-      end
-
-      def coord_to_decimal(point)
-        regex = /(?<dir>[NESW])\s*(?<deg>\d+)[°⁰º](?:(?<min>\d+)[ʹ'])?(?:(?<sec>\d+)[ʺ"])?/
-        match = regex.match(point)
-
-        return Float::INFINITY unless match
-
-        dec = match['deg'].to_i
-        dec += match['min'].to_f / 60
-        dec += match['sec'].to_f / 60 / 60
-        dec = -1 * dec if match['dir'] == 'W' || match['dir'] == 'S'
-
-        dec
+        cleaner_coordinate(value)
       end
     end
   end
